Refactored Table body to a separate component for hopefully better performance, added loading animation, fixed ordering
