--- packages/website/src/routes/Homepage/ReefTable/types.ts
@@ -1,3 +1,0 @@
-export type Order = "asc" | "desc" | undefined;
-
-export type OrderKeys = "locationName" | "temp" | "depth" | "dhw";
